tesseract-api/requirements.txt
fastapi
uvicorn
pillow
pytesseract
requests
python-multipart

